Prevent unnecessary branches for Put codepath
Currently, we loop over all processes and check to see if we have
any communication with each of them. This could be wasteful
since a process could be communicating with a few of the processes.

This commit prepares a list of processes that it needs to
communicate with before any communication. Hence, there are no
wasteful branches during communication.

The scope of this commit is for Put-based communication only.